Export SelectBooks so deprecations have a replacement
Follow-up to the PR review feedback asking that the deprecation notes
point users to a replacement. There was no non-deprecated, exported
book selector to recommend, so this exports one.

- Rename the scope selector's internal BookSelector to SelectBooks, and
  extract its combobox into a standalone SelectBooksPicker; export both
  from platform-bible-react
- Update the select-books dialog and the deprecated BookSelector
  deprecation notes to recommend SelectBooks as the replacement
