+++ lib/platform-bible-react/src/components/advanced/book-selector.stories.tsx
@@ -12,7 +12,7 @@ const meta: Meta<typeof BookSelector> = {
     docs: {
       description: {
         component: `
-> **Deprecated** (Jul 18 2025) — This component is no longer supported or tested. Use of this component is discouraged and it may be removed in the future.
+> **Deprecated** (Jul 18 2025) — This component is no longer supported or tested and will be removed in a later version. To let users select books, use the \`SelectBooks\` component instead (or \`ScopeSelector\` to combine scope and book selection).
 
 A UI for selecting books and chapter ranges. The user can either use the current book (with a start/end chapter range) or choose multiple books from a picker.
         `,
