+++ lib/platform-bible-react/src/components/advanced/scope-selector/scope-selector.utils.ts
@@ -1,4 +1,25 @@
+import { Canon } from '@sillsdev/scripture';
 import { getSectionForBook, Section } from 'platform-bible-utils';
+
+/**
+ * Derives the list of available, non-obsolete book IDs from the `availableBookInfo` string
+ *
+ * @param availableBookInfo Information about available books, formatted as a 123 character long
+ *   string as defined in a project's BooksPresent setting. Index N is '1' when the Nth canon book
+ *   is available
+ * @returns Array of available, non-obsolete book IDs in canon order
+ * @throws If `availableBookInfo` length does not match the number of canon books
+ */
+export const getAvailableBookIds = (availableBookInfo: string): string[] => {
+  if (availableBookInfo.length !== Canon.allBookIds.length) {
+    throw new Error('availableBookInfo length must match Canon.allBookIds length');
+  }
+
+  return Canon.allBookIds.filter(
+    (bookId, index) =>
+      availableBookInfo[index] === '1' && !Canon.isObsolete(Canon.bookIdToNumber(bookId)),
+  );
+};
 
 /**
  * Filters an array of book IDs to only include books from a specific section
